MultiBarGraphViews now nicely support multi-series bar graphing.
New stuff:
Fields. These replace horizontal labels in the MultiBarGraphView. They
keep track of their position so that the corresponding data fields can be
drawn centered over the field label in the graph.

PreDraw. This is for performance optimization. Its only really being
utilized in MultiBarGraphView right now.

Notes:
Overall, this MultiBarGraphView is an ugly hybrid between simply taking
advantage of the GraphView API and refactoring the GraphView API. There
really should be a quality refactoring.

Issues:
MultiBarGraphViews do not support zooming.
GraphViewData is now copied instead of referenced, meaning you can't
update this view live and easily (you'd have to reset all the data, which
isn't very effecient if everything is to be recopied).